BTFS-2340: Re-organize storage command files and folders
