+++ core/commands/restart.go
@@ -5,10 +5,10 @@ import (
 	"os/exec"
 	"time"
 
+	"github.com/TRON-US/go-btfs/core/commands/storage/path"
+
 	"github.com/TRON-US/go-btfs-cmds"
 	"github.com/cenkalti/backoff/v4"
-
-	path "github.com/TRON-US/go-btfs/core/commands/storage"
 )
 
 var daemonStartup = func() *backoff.ExponentialBackOff {
