+++ core/commands/storage/announce/announce.go
@@ -1,4 +1,4 @@
-package storage
+package announce
 
 import (
 	"fmt"
