+++ core/commands/storage/contracts/contracts.go
@@ -1,4 +1,4 @@
-package upload
+package contracts
 
 import (
 	"context"
