+++ core/commands/storage/hosts/hosts.go
@@ -1,4 +1,4 @@
-package storage
+package hosts
 
 import (
 	"context"
