+++ core/commands/storage/info/info.go
@@ -1,4 +1,4 @@
-package storage
+package info
 
 import (
 	"fmt"
