+++ core/commands/storage/path/path.go
@@ -1,4 +1,4 @@
-package storage
+package path
 
 import (
 	"errors"
