+++ core/commands/storage/stats/stats_test.go
@@ -1,4 +1,4 @@
-package storage
+package stats
 
 import (
 	"bytes"
