+++ core/commands/storage/store.go
@@ -1,7 +1,13 @@
 package storage
 
 import (
+	"github.com/TRON-US/go-btfs/core/commands/storage/announce"
 	"github.com/TRON-US/go-btfs/core/commands/storage/challenge"
+	"github.com/TRON-US/go-btfs/core/commands/storage/contracts"
+	"github.com/TRON-US/go-btfs/core/commands/storage/hosts"
+	"github.com/TRON-US/go-btfs/core/commands/storage/info"
+	"github.com/TRON-US/go-btfs/core/commands/storage/path"
+	"github.com/TRON-US/go-btfs/core/commands/storage/stats"
 	"github.com/TRON-US/go-btfs/core/commands/storage/upload/upload"
 
 	cmds "github.com/TRON-US/go-btfs-cmds"
@@ -16,12 +22,12 @@ host information sync/display operations, and BTT payment-related routines.`,
 	},
 	Subcommands: map[string]*cmds.Command{
 		"upload":    upload.StorageUploadCmd,
-		"hosts":     StorageHostsCmd,
-		"info":      StorageInfoCmd,
-		"announce":  StorageAnnounceCmd,
+		"hosts":     hosts.StorageHostsCmd,
+		"info":      info.StorageInfoCmd,
+		"announce":  announce.StorageAnnounceCmd,
 		"challenge": challenge.StorageChallengeCmd,
-		"stats":     StorageStatsCmd,
-		"contracts": upload.StorageContractsCmd,
-		"path":      PathCmd,
+		"stats":     stats.StorageStatsCmd,
+		"contracts": contracts.StorageContractsCmd,
+		"path":      path.PathCmd,
 	},
 }
